修bug，C#端成功绘制图元

+++ CLITestsMyGL/GLSMesh.cpp
@@ -4,11 +4,11 @@ GLSMesh::GLSMesh(array<float>^ vertices, array<unsigned int>^ indices)
 {
 	std::vector<float> verticesRaw;
 	std::vector<unsigned int> indicesRaw;
-	for each (int vertex in vertices)
+	for each (float vertex in vertices)
 	{
 		verticesRaw.push_back(vertex);
 	}
-	for each (int index in indices)
+	for each (unsigned int index in indices)
 	{
 		indicesRaw.push_back(index);
 	}
